web-server/requirements.txt
Flask
mysql-connector-python
python-dotenv
psutil